Optimise mem usage, disallowing random ref order
Currently, pycoverm is not particularly memory efficient. It keeps coverages
in an name=>Vec<f32> IndexMap, then creates a matrix and fills the matrxi from
the map.
This allows the order of reference sequences in the BAM files to be in different
order, but it consumes a lot of memory. Further, it's not very common to have a
situation where BAM files are mapped to the exact same set of references, but
in a different order, as BAM reference order is deterministic, and reflects the
order of the reference FASTA file.

This commit checks that all BAM files have the same headers in the same order.
It then computes an old_refindex => new_refindex map, which can be stored as a
single Vec<u32>. The coverage matrix can then be preallocated and filled in.
This approach is nearly perfectly memory efficient.